Algumas mudanças, porem com erros
