--- tenis-quarkus/src/main/java/br/unitins/joaovittor/basqueteiros/Cliente/dto/ClienteSaldoUpdateDTO.java
@@ -1,7 +1,0 @@
-package br.unitins.joaovittor.basqueteiros.Cliente.dto;
-
-public record ClienteSaldoUpdateDTO(
-    Double acrescimoSaldo
-) {
-
-}
