+++ tenis-quarkus/src/main/java/br/unitins/joaovittor/basqueteiros/Cliente/model/Cliente.java
@@ -1,10 +1,12 @@
 package br.unitins.joaovittor.basqueteiros.Cliente.model;
 
+import java.time.LocalDate;
 import java.util.List;
 
 import br.unitins.joaovittor.basqueteiros.DefaultEntity.model.DefaultEntity;
 import br.unitins.joaovittor.basqueteiros.Endereco.model.Endereco;
-import br.unitins.joaovittor.basqueteiros.PessoaFisica.model.PessoaFisica;
+import br.unitins.joaovittor.basqueteiros.Telefone.model.Telefone;
+import br.unitins.joaovittor.basqueteiros.Usuario.model.Usuario;
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -15,32 +17,26 @@ import jakarta.persistence.OneToOne;
 @Entity
 public class Cliente extends DefaultEntity{
 
-    @Column(name = "saldo")
-    private Double saldo;
-
     @OneToMany(cascade = CascadeType.ALL, orphanRemoval = true)
     @JoinColumn(name = "id_cliente")
     private List<Endereco> listaEndereco;
 
+    @Column(name = "cpf", unique = true, nullable = false)
+    private String cpf;
+
+    @Column(name = "nome")
+    private String nome;
+
+    @Column(name = "telefone")
+    private Telefone telefone;
+
+    @Column(name = "data_nascimento")
+    private LocalDate dataNascimento;
+
     @OneToOne
-    @JoinColumn(name = "id_pessoa_fisica", unique = true)
-    private PessoaFisica pessoaFisica;
+    @JoinColumn(name = "id_usuario", unique = true)
+    private Usuario usuario;
 
-    public Double getSaldo() {
-        return saldo;
-    }
-
-    public void setSaldo(Double saldo) {
-        this.saldo = saldo;
-    }
-
-    public PessoaFisica getPessoaFisica() {
-        return pessoaFisica;
-    }
-
-    public void setPessoaFisica(PessoaFisica pessoaFisica) {
-        this.pessoaFisica = pessoaFisica;
-    }
 
     public List<Endereco> getListaEndereco() {
         return listaEndereco;
@@ -50,6 +46,44 @@ public class Cliente extends DefaultEntity{
         this.listaEndereco = listaEndereco;
     }
 
-    
-        
+    public String getCpf() {
+        return cpf;
+    }
+
+    public void setCpf(String cpf) {
+        this.cpf = cpf;
+    }
+
+    public String getNome() {
+        return nome;
+    }
+
+    public void setNome(String nome) {
+        this.nome = nome;
+    }
+
+    public Telefone getTelefone() {
+        return telefone;
+    }
+
+    public void setTelefone(Telefone telefone) {
+        this.telefone = telefone;
+    }
+
+    public LocalDate getDataNascimento() {
+        return dataNascimento;
+    }
+
+    public void setDataNascimento(LocalDate dataNascimento) {
+        this.dataNascimento = dataNascimento;
+    }
+
+    public Usuario getUsuario() {
+        return usuario;
+    }
+
+    public void setUsuario(Usuario usuario) {
+        this.usuario = usuario;
+    }
+
 }
