--- tenis-quarkus/src/main/java/br/unitins/joaovittor/basqueteiros/Pessoa/model/Pessoa.java
@@ -1,47 +1,0 @@
-package br.unitins.joaovittor.basqueteiros.Pessoa.model;
-
-import java.time.LocalDate;
-
-import br.unitins.joaovittor.basqueteiros.DefaultEntity.model.DefaultEntity;
-import jakarta.persistence.Column;
-import jakarta.persistence.MappedSuperclass;
-
-@MappedSuperclass
-public class Pessoa extends DefaultEntity{
-    
-    @Column(name = "nome")
-    private String nome;
-
-    @Column(name = "telefone")
-    private String telefone;
-
-    @Column(name = "data_nascimento")
-    private LocalDate dataNascimento;
-
-    public String getNome() {
-        return nome;
-    }
-
-    public void setNome(String nome) {
-        this.nome = nome;
-    }
-
-    public String getTelefone() {
-        return telefone;
-    }
-
-    public void setTelefone(String telefone) {
-        this.telefone = telefone;
-    }
-
-    public LocalDate getDataNascimento() {
-        return dataNascimento;
-    }
-
-    public void setDataNascimento(LocalDate dataNascimento) {
-        this.dataNascimento = dataNascimento;
-    }
-
-    
-    
-}
