+++ tenis-quarkus/src/main/java/br/unitins/joaovittor/basqueteiros/PessoaFisica/model/PessoaFisica.java
@@ -10,8 +10,7 @@ import jakarta.persistence.OneToOne;
 @Entity
 public class PessoaFisica extends Pessoa{
     
-    @Column(name = "cpf", unique = true, nullable = false)
-    private String cpf;
+    
 
     @OneToOne
     @JoinColumn(name = "id_usuario", unique = true)
@@ -24,13 +23,4 @@ public class PessoaFisica extends Pessoa{
     public void setUsuario(Usuario usuario) {
         this.usuario = usuario;
     }
-
-    public String getCpf() {
-        return cpf;
-    }
-
-    public void setCpf(String cpf) {
-        this.cpf = cpf;
-    }
-    
 }
